Add roles: IntelMQ, Beats (elk-beats) & fix roles: Elk
The 'elk' role creates only the most basic installation, where Elasticsearch is configured along with a Logstash connector initialized to listen for a single json filebeat input.

The 'elk-beats' role installs Filebeat configured to process JSON input received on the localhost port and forward data to either Elasticsearch or Logstash.

The 'intelmq' role sets up a blank intelmq service along with a web-based GUI IntelQM Manager.

The 'redis' role sets up a basic local redis server running on default port.

Minor cleanups, including a separation of install step for redis-server as a separate role.

Added logging for Apache, IntelMQ and redis. The cronjob for logrotate is not created as part of the role.

Tested on Ubuntu server 16.04